veyron_rialto: Turn on all leds
Without this, the leds would be stuck to whatever the pullup/down states the
pins come with on rk3288.

Ready2_LED, an orange led, is one of the leds in this state.
This might confuse some users thinking there's an error.

Turn all of them on instead.
Later on depthcharge will use the same LEDs to indicate dev mode status.

BUG=chrome-os-partner:44274
BRANCH=master
TEST=Boot firmware without anything else, note all leds on

Change-Id: I5cf19aabd2a59a61699ef491ae11424cf5a0c874
Signed-off-by: Patrick Georgi <patrick@georgi-clan.de>
Original-Commit-Id: 2e1a332a5653fb76bbf8fe624274ec64d2b443a5
Original-Change-Id: I4c4e8940dd9cf1ac0301ac00bfc5992ba16e1589
Original-Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/294065
Original-Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: http://review.coreboot.org/11398
Tested-by: build bot (Jenkins)
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>